feat: implement get initial profile data
